requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --no-emit-index-url --no-emit-trusted-host --output-file=requirements.txt requirements/base.txt
#
asgiref==3.5.2
    # via django
django==4.1.2
    # via
    #   -r requirements/base.txt
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework==3.14.0
    # via
    #   -r requirements/base.txt
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via -r requirements/base.txt
psycopg2-binary==2.9.4
    # via -r requirements/base.txt
pyjwt==2.6.0
    # via djangorestframework-simplejwt
pytz==2022.5
    # via djangorestframework
sqlparse==0.4.3
    # via django
